+++ docs/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - ??
+
+### Added
+
+- `rsconnect content get-lockfile` command allows fetching a lockfile with the
+  dependencies installed by connect to run the deployed content
+- `rsconnect content venv` command recreates a local python environment
+  equal to the one used by connect to run the content.
+
+
 ## [1.28.2] - 2025-12-05
 
 ### Fixed